pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "protocolo-ouroboros"
version = "1.0.1"
description = "Pipeline ETL financeiro pessoal"
requires-python = ">=3.11"
license = {text = "GPL-3.0"}
authors = [{name = "AndreBFarias"}]

dependencies = [
    "pandas>=2.0",
    "openpyxl>=3.1",
    "xlrd>=2.0",
    "pdfplumber>=0.10",
    "Pillow>=10.0",
    "pytesseract>=0.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "msoffcrypto-tool>=5.0",
    "ofxparse>=0.21",
    "pikepdf>=8.0",
    "pillow-heif>=0.16",
]

keywords = ["etl", "finance", "personal", "streamlit", "dashboard", "irpf"]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial",
]

[project.urls]
Homepage = "https://github.com/AndreBFarias/protocolo-ouroboros"
Repository = "https://github.com/AndreBFarias/protocolo-ouroboros"
"Bug Tracker" = "https://github.com/AndreBFarias/protocolo-ouroboros/issues"
Changelog = "https://github.com/AndreBFarias/protocolo-ouroboros/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dashboard = [
    "streamlit>=1.30",
    "plotly>=5.18",
]
integrations = [
    "belvo-python>=0.35",
    "google-api-python-client>=2.0",
    "google-auth-oauthlib>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.3",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

# "A simplicidade é a sofisticação suprema." -- Leonardo da Vinci
